more cleanup

+++ AdventOfCode2023/Day3/SchematicValue.cs
@@ -5,7 +5,7 @@ namespace AdventOfCode2023.Day3;
 public class SchematicValue
 {
     public List<ValuePosition<char>> ValuePositions { get; set; } = new();
-    public int Value { get; set; } = 0;
+    public long Value { get; set; } = 0;
     public ValuePosition<char>? Symbol { get; set; } = null!;
     public bool IsPartNumber => Symbol != null;
